Move TensorRT utils in separate file, Minor clean up
